This program showcases the input method.  I took it a step further giving the user an option to continue printing to the console.  Error checking makes sure that a Y and or an N is given by the user and if not, to prompt the user to input a Y or an N.
I learned that you can change the input given by the user directly on the input path by using .upper() at the end of the input method.  You can change the case directly with the input method.  Also have practice with boolean variables of changing from true to false.  Used a number of while loops and if checks.